Airflow/requirements.txt
psycopg2-binary==2.9.1
cassandra-driver==3.29.0
